% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortedfreq.R
\name{sortedfreq}
\alias{sortedfreq}
\title{Sorted frequency}
\usage{
sortedfreq(df, colname)
}
\arguments{
\item{df}{The name of your dataframe}

\item{colname}{The column name to calculate sorted frequency.}
}
\description{
This function determines the frequency of a set of values. Returns the values and their frequencies. Values are sorted descending by frequency
}
\details{
These are the details
}
\examples{
sortedfreq(mtcars, "carb")
}
\keyword{frequency}
\keyword{sorted}
